creates line class to deal with rows, columns, diagonals and if there is marked line
